rename consultation symptom to encounter symptom
