+++ care/facility/models/__init__.py
@@ -4,8 +4,8 @@ from .patient_base import *  # noqa  # isort:skip
 from .ambulance import *  # noqa
 from .asset import *  # noqa
 from .bed import *  # noqa
-from .consultation_symptom import *  # noqa
 from .daily_round import *  # noqa
+from .encounter_symptom import *  # noqa
 from .events import *  # noqa
 from .facility import *  # noqa
 from .icd11_diagnosis import *  # noqa
