+++ care/facility/utils/reports/discharge_summary.py
@@ -12,9 +12,9 @@ from django.utils import timezone
 from hardcopy import bytestring_to_pdf
 
 from care.facility.models import (
-    ConsultationSymptom,
     DailyRound,
     Disease,
+    EncounterSymptom,
     InvestigationValue,
     PatientConsultation,
     PatientSample,
@@ -22,7 +22,7 @@ from care.facility.models import (
     PrescriptionDosageType,
     PrescriptionType,
 )
-from care.facility.models.consultation_symptom import ClinicalImpressionStatus
+from care.facility.models.encounter_symptom import ClinicalImpressionStatus
 from care.facility.models.file_upload import FileUpload
 from care.facility.models.icd11_diagnosis import (
     ACTIVE_CONDITION_VERIFICATION_STATUSES,
@@ -99,7 +99,7 @@ def get_discharge_summary_data(consultation: PatientConsultation):
     )
     hcx = Policy.objects.filter(patient=consultation.patient)
     daily_rounds = DailyRound.objects.filter(consultation=consultation)
-    symptoms = ConsultationSymptom.objects.filter(consultation=consultation).exclude(
+    symptoms = EncounterSymptom.objects.filter(consultation=consultation).exclude(
         clinical_impression_status=ClinicalImpressionStatus.ENTERED_IN_ERROR
     )
     diagnoses = get_diagnoses_data(consultation)
